refactor: umask is not useful for mkstemp
I misunderstood the utility of umask. mkstemp
creates the temporary file with permissions 600,
the permission level I'm looking for in the first
place, but that level cannot be altered before the
file is created anyway. This means that I not
only don't need umask for what I thought I did
but it wasn't doing what I intended anyway.