refactor: modularize functionality into separate files
As of now, the entire functionality of the project
is spread across a driver file and one .c and .h
file pair, even though there are distinct
groupings of functionality that could or should be
separated out into their own modules, e.g. our set
functionality. This commit modularizes the
functionality of the program, and makes other
changes such as the implementation of "module
private" functions through the use of the static
keyword and defining functions inside of their
associated .c file.