refactor!: revert part of a7931ac624de06e095115c3592e49689a359fa57
It is possible for this entire program to check
the number of our command line args once and be
fine, that I can tell. We never increase the
number of arguments, but do decrease them whenever
there are duplicates. While this is shorter, it
means that the set_init function has no guard on
it, though it has the ability to allocate
memory for our set struct, which does have limits
baked right into the struct def.

The consequence is that, if set_init is used
independent of our main function, such as in
testing, there is no way to protect against
overflow in the struct's key buffer.