fix: infinite loop when argc == 1
Problem:
mmv will freeze when given only one src argument

Reason:
When given only one argument, mmv will get stuck
in an infinite loop when it tries to detect
duplicates during cycle removal. This is because
the one src argument will be placed into the first
and only spot in the map. The map will only have
one spot because our formula for map capacity is
currently `(6 * (argc - 1)) + 1` == (6 * 0) + 1
for one argument. With only one space in the map,
the function for checking duplicates will attempt
to iterate with

`*hash = (*hash + 1 < set->map_capacity) ? *hash + 1 : 0;`

taking us back to the first position in the map
array forever.

Solution:
Only attempt to remove duplicates if there is
more than one src argument. This feels a little
bit like a kludge. Maybe a robust implementation
of the function to check for duplicates should
track whether it has traversed the entire
array.