feat: deduplicate absolute and relative paths
The user may avoid path deduplication by passing absolute paths to
mmv. For example, if the user gives a path to a relative file and an
absolute path to the same file, both will appear in the editing buffer,
allowing for them to conduct dangerous naming operations.

This commit initializes deduplication for this edge case. If the user
gives the "resolve paths" option, they will receive absolute paths
back. If they do not, they will receive the paths the entered as is,
but with duplicates removed. This allows for a mix of relative and
absolute paths.